design is 50% done

+++ app/src/main/java/com/example/weather/weather_forecasting_current_day/Winds.java
@@ -4,25 +4,25 @@ import com.google.gson.annotations.SerializedName;
 
 public class Winds {
     @SerializedName("speed")
-    private long speed;
+    private double speed;
     @SerializedName("deg")
-    private long deg;
+    private int deg;
     @SerializedName("gust")
     private double gust;
 
-    public long getSpeed() {
+    public double getSpeed() {
         return speed;
     }
 
-    public void setSpeed(long speed) {
+    public void setSpeed(double speed) {
         this.speed = speed;
     }
 
-    public long getDeg() {
+    public int getDeg() {
         return deg;
     }
 
-    public void setDeg(long deg) {
+    public void setDeg(int deg) {
         this.deg = deg;
     }
